* App: Compiler warning cleanup

* Gui: Compiler warning cleanup

* Assembly: Compiler warning cleanup

* Measure: Compiler warning cleanup

* Sketcher: Compiler warning cleanup

* TechDraw: Compiler warning cleanup

* PartDesign: Compiler warning cleanup

+++ src/Gui/Dialogs/DlgExpressionInput.cpp
@@ -1058,7 +1058,8 @@ void DlgExpressionInput::setMsgText()
     // elide text if it is going out of widget bounds
     // note: this is only 'rough elide', as this text is usually not very long;
     const int msgLinesLimit = 3;
-    if (wrappedMsg.size() > msgContentWidth / msgFontMetrics.averageCharWidth() * msgLinesLimit) {
+    if (static_cast<int>(wrappedMsg.size())
+        > msgContentWidth / msgFontMetrics.averageCharWidth() * msgLinesLimit) {
         const QString elidedMsg = msgFontMetrics.elidedText(
             QString::fromStdString(wrappedMsg),
             Qt::ElideRight,
